добавляет constants.js
