+++ js/data.js
@@ -1,33 +1,5 @@
 import {getRandomInteger} from './util.js';
-const POST_COUNT = 25;
-const NAMES = [
-  "Наташа",
-  "Алексей",
-  "Лиза",
-  "Мирослава",
-  "Дмитрий",
-  "Данила",
-  "Софа",
-  "Марьяна"
-];
-
-const MESSAGES = [
-  "Всё отлично!",
-  "В целом всё неплохо. Но не всё.",
-  "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
-  "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
-  "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
-  "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!"
-];
-
-const DESCRIPTIONS = [
-  "на чилле, на раслабоне!",
-  "Анапа 2024",
-  "Дача, шашлыки, что еще нужно для счастья?",
-  "Not pickme",
-  "В качалке с друзьями",
-  "Гламур"
-];
+import {POST_COUNT,NAMES,MESSAGES,DESCRIPTIONS} from'./constants.js';
 
 const createComment = () => {
   const randomNameIndex = getRandomInteger(0, NAMES.length - 1);
